Ratings * RateCounter
